backend/requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.9
requests==2.31.0
PyJWT==2.8.0